requirements.txt
asgiref==3.7.2
boto3==1.34.7
botocore==1.34.7
certifi==2023.11.17
charset-normalizer==3.3.2
cloudinary==1.38.0
dj-database-url==2.1.0
Django==4.2.5
django-cloudinary-storage==0.3.0
django-cors-headers==4.2.0
django-environ==0.11.2
django-filter==23.3
django-storages==1.14.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
gunicorn==21.2.0
idna==3.6
jmespath==1.0.1
packaging==23.2
Pillow==10.1.0
psycopg2==2.9.9
psycopg2-binary==2.9.9
PyJWT==2.8.0
python-dateutil==2.8.2
pytz==2023.3.post1
requests==2.31.0
s3transfer==0.10.0
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.9.0
tzdata==2023.3
